perf(tests): split database-free scope cases into their own file
The eight SimpleTestCase classes in posthog/test/test_scopes.py assert pure
scope transformations over module constants. They sat next to two BaseTest
classes, so the whole file needed Django database setup to run.

Move them to posthog/test/test_scopes_pure.py unchanged. Test bodies are
identical; only the imports and a placement comment are new.

Generated-By: PostHog Desktop
Task-Id: 82e4be60-6ce7-4818-ba40-40c0c9ea4d3f
